Fixed zenstore tests

+++ src/zenml/zen_stores/sql_zen_store.py
@@ -2197,6 +2197,9 @@ class SqlZenStore(BaseZenStore):
                     UserRoleAssignmentSchema.id == user_role_assignment_id
                 )
             ).one_or_none()
+            if not user_role:
+                raise KeyError(f"No user role assignment with id "
+                               f"{user_role_assignment_id} exists.")
 
             session.delete(user_role)
 
@@ -2299,6 +2302,9 @@ class SqlZenStore(BaseZenStore):
                     TeamRoleAssignmentSchema.id == team_role_assignment_id
                 )
             ).one_or_none()
+            if not team_role:
+                raise KeyError(f"No team role assignment with id "
+                               f"{team_role_assignment_id} exists.")
 
             session.delete(team_role)
